+++ storm-core/src/jvm/org/apache/storm/windowing/WatermarkTimeEvictionPolicy.java
@@ -22,7 +22,7 @@ package org.apache.storm.windowing;
  * watermark time and event lag into account.
  */
 public class WatermarkTimeEvictionPolicy<T> extends TimeEvictionPolicy<T> {
-    private final int lag;
+    private final long lag;
 
     /**
      * Constructs a WatermarkTimeEvictionPolicy that evicts events older
@@ -30,7 +30,7 @@ public class WatermarkTimeEvictionPolicy<T> extends TimeEvictionPolicy<T> {
      *
      * @param windowLength the window length in milliseconds
      */
-    public WatermarkTimeEvictionPolicy(int windowLength) {
+    public WatermarkTimeEvictionPolicy(long windowLength) {
         this(windowLength, Integer.MAX_VALUE);
     }
 
@@ -43,7 +43,7 @@ public class WatermarkTimeEvictionPolicy<T> extends TimeEvictionPolicy<T> {
      * @param windowLength the window length in milliseconds
      * @param lag          the max event lag in milliseconds
      */
-    public WatermarkTimeEvictionPolicy(int windowLength, int lag) {
+    public WatermarkTimeEvictionPolicy(long windowLength, long lag) {
         super(windowLength);
         referenceTime = 0L;
         this.lag = lag;
